Emulate step.
