+++ pkg/proc/native/threads_linux.go
@@ -47,30 +47,6 @@ func (t *nativeThread) resumeWithSig(sig int) (err error) {
 	return
 }
 
-func (t *nativeThread) singleStep() (err error) {
-	for {
-		t.dbp.execPtraceFunc(func() { err = sys.PtraceSingleStep(t.ID) })
-		if err != nil {
-			return err
-		}
-		wpid, status, err := t.dbp.waitFast(t.ID)
-		if err != nil {
-			return err
-		}
-		if (status == nil || status.Exited()) && wpid == t.dbp.pid {
-			t.dbp.postExit()
-			rs := 0
-			if status != nil {
-				rs = status.ExitStatus()
-			}
-			return proc.ErrProcessExited{Pid: t.dbp.pid, Status: rs}
-		}
-		if wpid == t.ID && status.StopSignal() == sys.SIGTRAP {
-			return nil
-		}
-	}
-}
-
 func (t *nativeThread) Blocked() bool {
 	regs, err := t.Registers()
 	if err != nil {
