+++ pkg/proc/native/threads_single_step_linux.go
@@ -1,0 +1,32 @@
+// +build !arm
+
+package native
+
+import (
+	"github.com/go-delve/delve/pkg/proc"
+	sys "golang.org/x/sys/unix"
+)
+
+func (t *nativeThread) singleStep() (err error) {
+	for {
+		t.dbp.execPtraceFunc(func() { err = sys.PtraceSingleStep(t.ID) })
+		if err != nil {
+			return err
+		}
+		wpid, status, err := t.dbp.waitFast(t.ID)
+		if err != nil {
+			return err
+		}
+		if (status == nil || status.Exited()) && wpid == t.dbp.pid {
+			t.dbp.postExit()
+			rs := 0
+			if status != nil {
+				rs = status.ExitStatus()
+			}
+			return proc.ErrProcessExited{Pid: t.dbp.pid, Status: rs}
+		}
+		if wpid == t.ID && status.StopSignal() == sys.SIGTRAP {
+			return nil
+		}
+	}
+}
